refactor: adding custom hook to separate the filter and sort logic
